backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
requests
fbprophet
python-dotenv